requirements.txt
rerun-sdk
sapien
ipython
setuptools
mani_skill
torch
torchvision 
torchaudio
